requirements.txt
asgiref==3.8.1
asttokens==3.0.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.1
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.6
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
executing==2.2.0
idna==3.10
ipython==8.32.0
jedi==0.19.2
matplotlib-inline==0.1.7
oauthlib==3.2.2
parso==0.8.4
prompt_toolkit==3.0.50
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
